pyproject.toml
[tool.poetry]

name = "afff4b45127463e9bb54d69246d7302"
version = "0.1.0"
description = ""
authors = ["Samuli Suomi <3490745+samulisuomi@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.2.1"
requests = "^2.28.1"
Flask-Cors = "^3.0.10"
python-dotenv = "^0.20.0"
redis = "^4.3.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
